% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example dataset for analysis}
\format{
A data frame with 146 observations and n avriables:
\describe{
\item{age}{Age in years}
\item{sex}{Sex ("Male", "Female")}
\item{smoker}{Smoking status ("Yes", "No")}
}
}
\usage{
data(data)
}
\description{
A sample dataset with demographic variables and smoking status of MS patients.
}
\keyword{datasets}
