% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plsda_clean.R
\name{run_plsda_clean}
\alias{run_plsda_clean}
\title{run_plsda_clean}
\usage{
run_plsda_clean(data, group, ncomp = 3, var.cutoff = 0.6)
}
\arguments{
\item{data}{A dataframe containing numeric and possibly non-numeric variables. If not, classify the variables according to data_for_plsda function above}

\item{group}{A factor or a vector that can be coerced to a factor, representing class labels.}

\item{ncomp}{Number of PLS components to compute. Default is 4.}

\item{var.cutoff}{Cutoff for variable correlation between covariates in plotVar. Default is 0.6.}
}
\value{
A list containing the plsda model and the generated plots.
}
\description{
\code{\link{run_plsda_clean}} Function for cleaning the dataset and running partial least squares discriminant analysis

This function separates categorical and numeric variables from a given dataframe
scales them, and runs PLS-DA using the mixOmics package.
It also plots the individual samples and  variables based on a cutoff of correlation.
}
\examples{
\dontrun{

library(mixOmics)

# Reading the data
data<- MSNeuroLab::data

 # Using the function for separating variables
 data.plsda <- data_for_plsda(data)
 
# Using the function for PLSDA
plsda <- run_plsda_clean(data = data, group=data$sex, ncomp = 2, var.cutoff = 0.6)

 }  


}
