% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_analysis.R
\name{similarity_analysis}
\alias{similarity_analysis}
\title{similarity_analysis}
\usage{
similarity_analysis(
  data = mtcars,
  input_columns = c("mpg", "cyl", "hp"),
  group_column = "am",
  patient_value = 1,
  control_value = 0,
  comparison_variable = "disp",
  distance_metric = "euclidean",
  x_label = NULL,
  y_label = NULL,
  plot_title = NULL,
  include_id = FALSE
)
}
\arguments{
\item{data}{data.frame that contains numeric or categorical variables arranged in columns.}

\item{input_columns}{List of columns that are used in computing distance metrics.}

\item{group_column}{Column containing categories for the two groups whose similarity in being analysed, for example MS patient and healthy control.}

\item{patient_value}{Value in group_column that identifies patients. Every patient is matched to one control. Several patients may match the same control.}

\item{control_value}{Value in group_column that identifies controls. Every control may not be match to any patient.}

\item{comparison_variable}{Compare this variable between patients and their most similar controls.}

\item{distance_metric}{Metric used to compute difference values. Choose "euclidean", "cosine", or "manhattan".}

\item{x_label}{Variable to plot on x-axis. Variable name is used if NULL.}

\item{y_label}{Variable to plot on y-axis. Variable name is used if NULL.}

\item{plot_title}{Plot title. Will be auto-generated if NULL.}

\item{include_id}{Whether to include patient/control IDs in results table.}
}
\value{
\item{result}{List of a table with numeric comparison_variable values, distance metrics and similarities for pairs of most similar patients and controls, and a scatterplot visualizing the comparisons}
}
\description{
\code{\link{similarity_analysis}} This function performs similarity matching between patients and controls based on multiple variables. It identifies the most similar control subject for each patient using distance-based similarity metrics.
}
\examples{
\dontrun{
# Using the function
test_results <- similarity_analysis() # By default generates an example plot from R test data set mtcars
print(test_results$results) # Table of comparison_variable values, distance metrics and similarities for pairs of most similar patients and controls 
print(test_results$plot) # Scatterplot showing correlation in comparison_variable between the most similar patients and controls with one dot for each patient-control comparison. 
# One dot will match each patient. Each control is matched by none to several dots. Dot size indicates level of similarity calculated from input columns. 
# The farther away from the dotted line the dot lands, the bigger the discrepancy in comparison_variable between the patient and the control.
 }                             

 @export
}
