% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_univariate_logistic.R
\name{run_univariate_logistic}
\alias{run_univariate_logistic}
\title{run_univariate_logistic}
\usage{
run_univariate_logistic(data, outcome_var, predictors)
}
\arguments{
\item{data}{A data frame containing the outcome and predictor variables.}

\item{outcome_var}{A string specifying the name of the binary outcome variable.
If the variable is a factor or character, values equal to "Yes" are
recoded to 1 and others to 0.}

\item{predictors}{A character vector specifying the names of the predictor variables to test.}
}
\value{
A data frame with the following columns:
\describe{
\item{Predictor}{Name of the predictor variable}
\item{OR}{Estimated odds ratio}
\item{CI_low}{Lower bound of the 95\% confidence interval for the OR}
\item{CI_high}{Upper bound of the 95\% confidence interval for the OR}
\item{p_value}{p-value for the predictor in the logistic model}
}
}
\description{
\code{\link{run_univariate_logistic}} This function performs univariate logistic regression analyses for a set of predictor variables against a binary outcome variable in a given dataset. It returns a tidy data frame with odds ratios (OR), 95\% confidence intervals (CI), and p-values for each predictor.
}
\examples{

# Reading the data
data<- MSNeuroLab::data
# Selecting variables to test at univariate 
variables_to_test <- colnames(data)[c(3:8)]
# Using the function
res_univariates <- run_univariate_logistic(data, "smoker", predictors = variables_to_test)




}
