% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_with_levene.R
\name{test_with_levene}
\alias{test_with_levene}
\title{test_with_levene}
\usage{
test_with_levene(data, group_var)
}
\arguments{
\item{data}{data.frame object with the variables (numeric or categorical) on which performing the statistical comparisons}

\item{group_var}{grouping variable with two groups (classes) in quotes}
}
\value{
\item{final_table}{a data.frame object with results unlisted (with bindrow function)}
}
\description{
\code{\link{test_with_levene}} Function for comparing statistical comparison between numeric variables, testing for assumptions on varaibles' distribution and categorical variables
}
\examples{
\dontrun{

library(readxl)

# Reading the data
data<- read_excel("Data/examination_dataset1.xlsx")
 
# Using the function
test_stat <- test_with_levene(data = data[, c(1:4)],  group_var="sex")

 }                             


 
 
 @export
}
